ADDED: Se agrego nuevos modelos

+++ api/models/pagos.py
@@ -6,10 +6,11 @@ from api.models import ModeloBase
 
 
 class Pago(ModeloBase):
+
     usuario = models.ForeignKey(
         'api.Usuario', on_delete=models.CASCADE, related_name='pagos')
     servicio = models.ForeignKey(
         'api.Servicio', on_delete=models.CASCADE, related_name='pagos')
-    mes = models.IntegerField()
-    anio = models.IntegerField()
+    mes = models.PositiveSmallIntegerField()
+    anio = models.PositiveSmallIntegerField()
     pago = models.FloatField()
